requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
